requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.0
ollama==0.6.1
pgvector==0.4.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2